backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.9.2
pydantic-settings==2.5.2
openpyxl==3.1.5
reportlab==4.2.5
python-multipart==0.0.20




